# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PromptFolio API (served by the Python FastAPI backend)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      parameters:
        - name: role
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AgentRole"
      responses:
        "200":
          description: Agents (system defaults + user agents)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    patch:
      operationId: updateAgent
      tags: [agents]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    delete:
      operationId: deleteAgent
      tags: [agents]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ---------------- BUILDER ----------------
  /builder/sessions:
    get:
      operationId: listBuilderSessions
      tags: [builder]
      parameters:
        - name: workspaceType
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/BuilderWorkspaceType"
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSession"
    post:
      operationId: createBuilderSession
      tags: [builder]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionInput"
      responses:
        "201":
          description: Created (agent not yet locked)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
  /builder/sessions/{id}:
    get:
      operationId: getBuilderSession
      tags: [builder]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Session with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSessionDetail"
    delete:
      operationId: deleteBuilderSession
      tags: [builder]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /builder/sessions/{id}/messages:
    post:
      operationId: sendBuilderMessage
      tags: [builder]
      summary: >
        Send a message. On the first message agentId is required and becomes the
        permanently locked agent. Later messages ignore agentId; a different agentId returns 409.
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: Exchange result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatExchange"
  /builder/sessions/{id}/draft:
    get:
      operationId: getPublishDraft
      tags: [builder]
      summary: Pre-filled values for the Save & Publish final review card
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishDraft"
  /builder/sessions/{id}/publish:
    post:
      operationId: publishFromSession
      tags: [builder]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishInput"
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResult"

  # ---------------- DELIVERY ----------------
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }

    # ---------- enums ----------
    Modality:
      type: string
      enum: [image, video, audio, text]
...
    AgentRole:
      type: string
      enum: [meta_agent_builder, prompt_builder, delivery_executor]
    BuilderWorkspaceType:
      type: string
      enum: [builder_agent, builder_prompt]
    McpProviderType:
      type: string
      enum: [google_nano_banana, openai_chatgpt, runway, custom_mcp]
...
    PromptVariable:
      type: object
      required: [name]
      properties:
        name: { type: string }
        label: { type: ["string", "null"] }
        defaultValue: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        helpText: { type: ["string", "null"] }
        examples:
          type: array
          items: { type: string }
        options:
          type: array
          items: { type: string }
        required: { type: boolean }
...
    AssetThumb:
      type: object
      required: [id, modality, previewUrl]
      properties:
        id: { type: string }
        modality: { $ref: "#/components/schemas/Modality" }
        previewUrl: { type: string }
    Prompt:
      type: object
      required: [id, title, promptTemplate, variables, tags, targetModality, isPublished, version, recentAssets, assetCount, createdAt, updatedAt]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: ["string", "null"] }
        promptTemplate: { type: string }
        variables:
          type: array
          items: { $ref: "#/components/schemas/PromptVariable" }
        tags:
          type: array
          items: { type: string }
        folder: { type: ["string", "null"] }
        categoryId: { type: ["string", "null"] }
        categoryName: { type: ["string", "null"] }
        categoryColor: { type: ["string", "null"] }
        targetModality: { $ref: "#/components/schemas/Modality" }
        preferredMcpId: { type: ["string", "null"] }
        preferredMcpName: { type: ["string", "null"] }
        sourceAgentName: { type: ["string", "null"] }
        isPublished: { type: boolean }
        version: { type: integer }
        recentAssets:
          type: array
          items: { $ref: "#/components/schemas/AssetThumb" }
        assetCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Agent:
      type: object
      required: [id, name, agentRole, systemPrompt, isSystemDefault, isPublished, isActive, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        agentRole: { $ref: "#/components/schemas/AgentRole" }
        systemPrompt: { type: string }
        categoryId: { type: ["string", "null"] }
        categoryName: { type: ["string", "null"] }
        preferredMcpId: { type: ["string", "null"] }
        temperature: { type: number }
        isSystemDefault: { type: boolean }
        isPublished: { type: boolean }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    AgentInput:
      type: object
      required: [name, agentRole, systemPrompt]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
        agentRole: { $ref: "#/components/schemas/AgentRole" }
        systemPrompt: { type: string, minLength: 1 }
        categoryId: { type: ["string", "null"] }
        preferredMcpId: { type: ["string", "null"] }
        temperature: { type: number }
    AgentUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
        systemPrompt: { type: string, minLength: 1 }
        categoryId: { type: ["string", "null"] }
        preferredMcpId: { type: ["string", "null"] }
        temperature: { type: number }
        isActive: { type: boolean }
        isPublished: { type: boolean }

    # ---------- builder ----------
    ChatSession:
      type: object
      required: [id, workspaceType, title, isLocked, messageCount, createdAt]
      properties:
        id: { type: string }
        workspaceType: { $ref: "#/components/schemas/BuilderWorkspaceType" }
        title: { type: string }
        lockedAgentId: { type: ["string", "null"] }
        lockedAgentName: { type: ["string", "null"] }
        isLocked: { type: boolean }
        messageCount: { type: integer }
        publishedPromptId: { type: ["string", "null"] }
        publishedAgentId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    ChatSessionInput:
      type: object
      required: [workspaceType]
      properties:
        workspaceType: { $ref: "#/components/schemas/BuilderWorkspaceType" }
        title: { type: string }
    ChatMessage:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id: { type: string }
        role:
          type: string
          enum: [user, assistant]
        content: { type: string }
        createdAt: { type: string, format: date-time }
    ChatSessionDetail:
      type: object
      required: [session, messages]
      properties:
        session: { $ref: "#/components/schemas/ChatSession" }
        lockedAgent:
          oneOf:
            - $ref: "#/components/schemas/Agent"
            - type: "null"
        messages:
          type: array
          items: { $ref: "#/components/schemas/ChatMessage" }
    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }
        agentId:
          type: ["string", "null"]
          description: Required only for the first message of a session
    ChatExchange:
      type: object
      required: [session, userMessage, assistantMessage]
      properties:
        session: { $ref: "#/components/schemas/ChatSession" }
        userMessage: { $ref: "#/components/schemas/ChatMessage" }
        assistantMessage: { $ref: "#/components/schemas/ChatMessage" }
    PublishDraft:
      type: object
      required: [kind, title, content, variables, tags, targetModality]
      properties:
        kind:
          type: string
          enum: [prompt, agent]
        title: { type: string }
        description: { type: ["string", "null"] }
        content:
          type: string
          description: Consolidated text (prompt template or agent system prompt)
        variables:
          type: array
          items: { $ref: "#/components/schemas/PromptVariable" }
        tags:
          type: array
          items: { type: string }
        targetModality: { $ref: "#/components/schemas/Modality" }
        suggestedMcpProvider:
          oneOf:
            - $ref: "#/components/schemas/McpProviderType"
            - type: "null"
    PublishInput:
      type: object
      required: [kind, title, content]
      properties:
        kind:
          type: string
          enum: [prompt, agent]
        title: { type: string, minLength: 1 }
        description: { type: ["string", "null"] }
        content: { type: string, minLength: 1 }
        categoryId: { type: ["string", "null"] }
        folder: { type: ["string", "null"] }
        tags:
          type: array
          items: { type: string }
        variables:
          type: array
          items: { $ref: "#/components/schemas/PromptVariable" }
        targetModality: { $ref: "#/components/schemas/Modality" }
        preferredMcpId: { type: ["string", "null"] }
        agentRole: { $ref: "#/components/schemas/AgentRole" }
    PublishResult:
      type: object
      required: [kind]
      properties:
        kind:
          type: string
          enum: [prompt, agent]
        prompt:
          oneOf:
            - $ref: "#/components/schemas/Prompt"
            - type: "null"
        agent:
          oneOf:
            - $ref: "#/components/schemas/Agent"
            - type: "null"

    # ---------- delivery ----------
...